+++ src/main/java/org/nazar/service/dao/ParsersLinkJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParsersLinkJpaRepository extends JpaRepository<ParsersLink, Integer> {
+public interface ParsersLinkJpaRepository extends JpaRepository<ParsersLink, Long> {
     Optional<ParsersLink> findByProviderName(String providerName);
 }
